+++ Examples/Algorithms/Digitization/src/DigitizationConfig.cpp
@@ -8,40 +8,7 @@
 
 #include "ActsExamples/Digitization/DigitizationConfig.hpp"
 
-#include "Acts/Definitions/TrackParametrization.hpp"
-#include "Acts/Geometry/GeometryIdentifier.hpp"
-#include "ActsExamples/Digitization/SmearingConfig.hpp"
-
 namespace ActsExamples {
-
-DigitizationConfig::DigitizationConfig() = default;
-
-DigitizationConfig::DigitizationConfig(
-    Acts::GeometryHierarchyMap<DigiComponentsConfig> digiCfgs)
-    : digitizationConfigs(std::move(digiCfgs)) {}
-
-std::vector<
-    std::pair<Acts::GeometryIdentifier, std::vector<Acts::BoundIndices>>>
-DigitizationConfig::getBoundIndices() const {
-  std::vector<
-      std::pair<Acts::GeometryIdentifier, std::vector<Acts::BoundIndices>>>
-      bIndexInput;
-
-  for (std::size_t ibi = 0; ibi < digitizationConfigs.size(); ++ibi) {
-    Acts::GeometryIdentifier geoID = digitizationConfigs.idAt(ibi);
-    const auto dCfg = digitizationConfigs.valueAt(ibi);
-    std::vector<Acts::BoundIndices> boundIndices;
-    boundIndices.insert(boundIndices.end(),
-                        dCfg.geometricDigiConfig.indices.begin(),
-                        dCfg.geometricDigiConfig.indices.end());
-    // we assume nobody will add multiple smearers to a single bound index
-    for (const auto& c : dCfg.smearingDigiConfig) {
-      boundIndices.push_back(c.index);
-    }
-    bIndexInput.push_back({geoID, boundIndices});
-  }
-  return bIndexInput;
-}
 
 std::vector<Acts::ActsScalar> GeometricConfig::variances(
     const std::array<std::size_t, 2u>& csizes,
